backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.3.0
openai==1.12.0
pinecone-client==3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
httpx==0.26.0
